# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Swagger Liked Videos",
      "description": "A sample API that stores Video Id's of liked videos and also retreives. swagger-2.0 specification",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
        "name": "Swagger API Team"
      },
      "license": {
        "name": "MIT"
      }
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
      "/user": {
        "get": {
          "description": "Returns all the data stored about videos",
          "operationId": "getVideos",
          "produces": [
            "application/json",
            "application/xml",
            "text/xml",
            "text/html"
          ],
         
          "responses": {
            "200": {
              "description": "videos response",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Video"
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "schema": {
                "$ref": "#/definitions/ErrorModel"
              }
            }
          }
        },
        "post": {
          "description": "Creates a new video entry in the table.  Duplicates are allowed",
          "operationId": "addVideos",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Video",
              "in": "body",
              "description": "Video to add to the store",
              "required": true,
              "schema": {
                "$ref": "#/definitions/NewVideo"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "video response",
              "schema": {
                "$ref": "#/definitions/Video"
              }
            },
            "default": {
              "description": "unexpected error",
              "schema": {
                "$ref": "#/definitions/ErrorModel"
              }
            }
          }
        }
      }
      
    },
    "definitions": {
      "Video": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/definitions/NewVideo"
          }
          
        ]
      },
      "NewVideo": {
        "type": "object",
        "required": [
          "videoId"
        ],
        "properties": {
          "videoId": {
            "type": "string"
          },
          "Liked": {
            "type": "string"
          }
        }
      },
      "ErrorModel": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }